Renamed trainer to model. Renamed main to main_train Main_trainer function working in main_train
